+++ agrest-sencha/src/main/java/io/agrest/sencha/AgProtocolSenchaExt.java
@@ -1,9 +1,0 @@
-package io.agrest.sencha;
-
-/**
- * @since 4.1
- */
-public enum AgProtocolSenchaExt {
-
-    query, filter, group, groupDir;
-}