+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #515 Switch to Java 11
 * #516 Upgrade jackson to 2.13.1 
+* #517 Remove "agrest-sencha"
 
 ## Release 4.9